+++ lib/mapper.dart
@@ -4,7 +4,7 @@ import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_database/firebase_database.dart';
-import '../helpers.dart';
+import 'src/helpers.dart';
 // -----------------------------------------------------------------------------
 part 'src/methods.dart';
 // -----------------------------------------------------------------------------