+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+* Added Ai generated method tests, but still not finished, 16 methods still not tested by ai but tested and works 
+  perfect by manual testing in bldrs app
+
+
 ## 0.0.1
 
-* Separated Mapper class from bldrs into this package
+* Separated Mapper class from bldrs app into this package